update path config in nginx

+++ vars/configureNginxAndCertbot.groovy
@@ -1,14 +1,15 @@
 def call(String subdomain, String domain, String deployPort) {
-    // Check if the variables are passed
+    // Check if the required variables are passed
     if (!subdomain || !domain || !deployPort) {
         error "subdomain, domain, and deployPort must be provided"
     }
 
-    // Echo the variables to check if they are passed down correctly
+    // Echo the variables for debugging purposes
     echo "Subdomain: ${subdomain}"
     echo "Domain: ${domain}"
     echo "Deploy Port: ${deployPort}"
 
+    // Trim variables and ensure they are not empty
     if (subdomain.trim() == "" || domain.trim() == "" || deployPort.trim() == "") {
         error "subdomain, domain, and deployPort cannot be empty"
     }
@@ -20,6 +21,11 @@ def call(String subdomain, String domain, String deployPort) {
     folder_name="${subdomain}.${domain}"
     file_path="/etc/nginx/sites-available/\${folder_name}"
 
+    # Check if Nginx config file already exists
+    if [ -f "\${file_path}" ]; then
+        echo "Nginx config for \${folder_name} already exists."
+        exit 0
+    fi
 
     # Write the Nginx configuration to the file
     sudo bash -c "cat > \${file_path} <<EOL
@@ -38,6 +44,21 @@ def call(String subdomain, String domain, String deployPort) {
     EOL"
 
     # Create a symlink in /etc/nginx/sites-enabled/ to enable the site
-    sudo ln -s \${file_path} /etc/nginx/sites-enabled/\${folder_name}
+    if [ ! -L /etc/nginx/sites-enabled/\${folder_name} ]; then
+        sudo ln -s \${file_path} /etc/nginx/sites-enabled/\${folder_name}
+    else
+        echo "Symlink already exists for \${folder_name}"
+    fi
+
+    # Test the Nginx configuration for syntax errors
+    sudo nginx -t
+    if [ $? -eq 0 ]; then
+        # Reload Nginx to apply the new configuration
+        sudo systemctl reload nginx
+        echo "Nginx configuration reloaded for \${folder_name}"
+    else
+        echo "Nginx configuration test failed"
+        exit 1
+    fi
     """
 }
